Going to floats in FlatTreeProducer

+++ SexaQAnalysis/AnalyzerAllSteps/interface/FlatTreeProducer.h
@@ -52,17 +52,17 @@ class FlatTreeProducer : public edm::EDAnalyzer
    
     TTree* _tree;   
 
-    //definition of variablew which should go to tree
-    std::vector<double> _S_charge;
-    std::vector<double> _S_lxy_interaction_vertex, _S_error_lxy_interaction_vertex,_S_mass,_S_chi2_ndof;
-    std::vector<double> _S_daughters_deltaphi,_S_daughters_deltaeta,_S_daughters_openingsangle,_S_eta;
-    std::vector<double> _S_dxy,_Ks_dxy,_Lambda_dxy;
-    std::vector<double> _S_dxy_over_lxy,_Ks_dxy_over_lxy,_Lambda_dxy_over_lxy;
-    std::vector<double> _S_dz,_Ks_dz,_Lambda_dz;
-    std::vector<double> _S_pt,_Ks_pt,_Lambda_pt;
-    std::vector<double> _S_pz,_Ks_pz,_Lambda_pz;
-    std::vector<double> _S_vz_interaction_vertex;
-    std::vector<double> _S_vx,_S_vy,_S_vz;
+    //definition of variables which should go to tree
+    std::vector<float> _S_charge;
+    std::vector<float> _S_lxy_interaction_vertex, _S_error_lxy_interaction_vertex,_S_mass,_S_chi2_ndof;
+    std::vector<float> _S_daughters_deltaphi,_S_daughters_deltaeta,_S_daughters_openingsangle,_S_eta;
+    std::vector<float> _S_dxy,_Ks_dxy,_Lambda_dxy;
+    std::vector<float> _S_dxy_over_lxy,_Ks_dxy_over_lxy,_Lambda_dxy_over_lxy;
+    std::vector<float> _S_dz,_Ks_dz,_Lambda_dz;
+    std::vector<float> _S_pt,_Ks_pt,_Lambda_pt;
+    std::vector<float> _S_pz,_Ks_pz,_Lambda_pz;
+    std::vector<float> _S_vz_interaction_vertex;
+    std::vector<float> _S_vx,_S_vy,_S_vz;
 
 
      };
